Refs #8613 Fix bug: don't show the folder groups in the Team tab
We were using plone.api to revoke roles, which caused some
unanticipated side effects. This should probably be fixed upstream but
for now we revoke roles by manipulating __ac_local_roles__ directly.

api.user.revoke roles was adding the Member local role for the user,
and when revoking the Contributor role it replaced the role with
Editor.

This change also simplifies the code somewhat, so there is no need to
use plone.api here.